+++ src/commands/install_catalog.py
