+++ tests/commands/test_command_download.py
@@ -17,7 +17,7 @@ if str(project_root) not in sys.path:
     sys.path.insert(0, str(project_root))
 
 # Import the modules directly to avoid import issues
-from src.commands.download import DownloadCommand
+from src.commands.install_url import DownloadCommand
 
 
 @pytest.fixture
@@ -59,7 +59,7 @@ def mocked_parser(monkeypatch: pytest.MonkeyPatch, download_test_data: tuple[str
     mock.ask_url.return_value = None
 
     parser_class_mock = MagicMock(return_value=mock)
-    monkeypatch.setattr("src.commands.download.ParseURL", parser_class_mock)
+    monkeypatch.setattr("src.commands.install_url.ParseURL", parser_class_mock)
     return mock
 
 
@@ -86,7 +86,7 @@ def mocked_app_config(
     mock.config_file_name = "test_config.json"
 
     app_config_class_mock = MagicMock(return_value=mock)
-    monkeypatch.setattr("src.commands.download.AppConfigManager", app_config_class_mock)
+    monkeypatch.setattr("src.commands.install_url.AppConfigManager", app_config_class_mock)
     return mock
 
 
@@ -114,7 +114,7 @@ def mocked_api(monkeypatch: pytest.MonkeyPatch, download_test_data: tuple[str, s
     mock.appimage_url = download_test_data["appimage_url"]
 
     api_class_mock = MagicMock(return_value=mock)
-    monkeypatch.setattr("src.commands.download.GitHubAPI", api_class_mock)
+    monkeypatch.setattr("src.commands.install_url.GitHubAPI", api_class_mock)
     return mock
 
 
@@ -133,7 +133,7 @@ def mocked_download_manager(monkeypatch: pytest.MonkeyPatch) -> MagicMock:
     mock.download.return_value = True
 
     download_manager_class_mock = MagicMock(return_value=mock)
-    monkeypatch.setattr("src.commands.download.DownloadManager", download_manager_class_mock)
+    monkeypatch.setattr("src.commands.install_url.DownloadManager", download_manager_class_mock)
     return mock
 
 
@@ -179,7 +179,7 @@ def mocked_verifier(
     mock.verify_appimage.return_value = verification_success
 
     verifier_class_mock = MagicMock(return_value=mock)
-    monkeypatch.setattr("src.commands.download.VerificationManager", verifier_class_mock)
+    monkeypatch.setattr("src.commands.install_url.VerificationManager", verifier_class_mock)
     return mock
 
 
@@ -198,7 +198,7 @@ def mocked_verifier_failure(monkeypatch: pytest.MonkeyPatch) -> MagicMock:
     mock.verify_appimage.return_value = False
 
     verifier_class_mock = MagicMock(return_value=mock)
-    monkeypatch.setattr("src.commands.download.VerificationManager", verifier_class_mock)
+    monkeypatch.setattr("src.commands.install_url.VerificationManager", verifier_class_mock)
     return mock
 
 
@@ -217,7 +217,7 @@ def mocked_file_handler(monkeypatch: pytest.MonkeyPatch) -> MagicMock:
     mock.handle_appimage_operations.return_value = True
 
     file_handler_class_mock = MagicMock(return_value=mock)
-    monkeypatch.setattr("src.commands.download.FileHandler", file_handler_class_mock)
+    monkeypatch.setattr("src.commands.install_url.FileHandler", file_handler_class_mock)
     return mock
 
 
@@ -235,7 +235,7 @@ def mocked_icon_manager(monkeypatch: pytest.MonkeyPatch) -> MagicMock:
     mock = MagicMock()
 
     icon_manager_class_mock = MagicMock(return_value=mock)
-    monkeypatch.setattr("src.commands.download.IconManager", icon_manager_class_mock)
+    monkeypatch.setattr("src.commands.install_url.IconManager", icon_manager_class_mock)
     return mock
